navbar and unfinished testing
